requirements.txt
Django==2.2
django-crispy-forms==1.7.2
Markdown==3.1
pytz==2021.3
sqlparse==0.4.2
xlwt==1.3.0
